+++ shared/sys/con_win32.cpp
@@ -138,7 +138,7 @@ static void CON_HistPrev( void )
 
 	qconsole_history_pos = pos;
 	Q_strncpyz( qconsole_line, qconsole_history[ qconsole_history_pos ], sizeof( qconsole_line ) );
-	qconsole_linelen = strlen( qconsole_line );
+	qconsole_linelen = static_cast<int>(strlen( qconsole_line ));
 	qconsole_cursor = qconsole_linelen;
 }
 
@@ -169,7 +169,7 @@ static void CON_HistNext( void )
 
 	qconsole_history_pos = pos;
 	Q_strncpyz( qconsole_line, qconsole_history[ qconsole_history_pos ], sizeof( qconsole_line ) );
-	qconsole_linelen = strlen( qconsole_line );
+	qconsole_linelen = static_cast<int>(strlen( qconsole_line ));
 	qconsole_cursor = qconsole_linelen;
 }
 
@@ -406,7 +406,7 @@ char *CON_Input( void )
 				Q_strncpyz( f.buffer, qconsole_line, sizeof( f.buffer ) );
 				Field_AutoComplete( &f );
 				Q_strncpyz( qconsole_line, f.buffer, sizeof( qconsole_line ) );
-				qconsole_linelen = strlen( qconsole_line );
+				qconsole_linelen = static_cast<int>(strlen( qconsole_line ));
 				qconsole_cursor = qconsole_linelen;
 
 				break;
